pyproject.toml
[project]
name = "email-automation-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.0",
    "fastapi>=0.116.1",
    "google-api-python-client>=2.176.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "requests>=2.32.4",
    "rich>=14.0.0",
]
